requirements.txt
Django==2.0.2
pkg-resources==0.0.0
pytz==2018.3
django-wkhtmltopdf-3.2.0
djangorestframework-3.8.2
pillow-5.2.0
